terraform: lang.Data fake for use in unit tests
Our real implementation of the lang.Data interface depends on just about
everything else in the modules runtime to implement its behavior, and so
it's not really appropriate to use in unit tests.

This new implementation is considerably simpler, just returning values
predefined in a set of fixed tables. This is appropriate for unit tests of
features that perform expression evaluation as a side-effect of their work
but that don't contribute their own values to the real evaluation data
implementation.